feat(desktop): move fullview plugin buttons from title bar to left rail
Relocates fullview plugin icons (e.g. Tasks) to the left rail, above
the Settings and About buttons with a divider separator. Aligns the
title bar right-side content with the right rail width.
